Fixes for building under MSVC.

+++ libdcpu-vm-dbg/dbglua.c
@@ -310,7 +310,7 @@ int dbg_lua_handle_break(lua_State* L)
 			ddbg_return_code = 1;
 	}
 	else if (lua_isnumber(L, 2))
-		ddbg_return_code = lua_tonumber(L, 2);
+		ddbg_return_code = (int)lua_tonumber(L, 2);
 	return 0;
 }
 
@@ -425,7 +425,7 @@ void dbg_lua_push_state(struct lua_debugst* ds, struct dbg_state* state, void* u
 	lua_setfield(ds->state, tbl, "lines");
 	lua_pushcfunction(ds->state, &dbg_lua_handle_raise);
 	lua_setfield(ds->state, tbl, "raise");
-	
+
 	// Push CPU state into table.
 	vm_hw_lua_cpu_push_cpu(ds->state, state->get_vm());
 	lua_setfield(ds->state, tbl, "cpu");
@@ -518,12 +518,12 @@ void dbg_lua_handle_command(struct dbg_state* state, void* ud, freed_bstring nam
 		lua_pop(ds->state, 2);
 		list_iterator_stop(&modules);
 		list_destroy(parameters);
-		
+
 		// The command may have started the virtual machine, check the
 		// status of the VM and execute if needed.
 		if (state->get_vm()->halted == false)
 			vm_execute(state->get_vm(), NULL);
-		
+
 		return;
 	}
